Make ol-dumps easier to test

+++ scripts/oldump.sh
@@ -28,6 +28,9 @@
 
 set -e
 
+# To run a testing subset of the full ol-dump, uncomment the following line.
+# export OLDUMP_TESTING=true
+
 if [ $# -lt 1 ]; then
     echo "USAGE: $0 yyyy-mm-dd [--archive]" 1>&2
     exit 1
@@ -52,8 +55,9 @@ function log() {
 
 # create a clean directory
 log "clean directory: $TMPDIR/dumps"
-rm -rf $TMPDIR/dumps  # TESTING: comment this out
 mkdir -p $TMPDIR/dumps
+# Remove any leftover ol_cdump* and ol_dump* files or directories.
+rm -rf $TMPDIR/dumps/ol_*
 cd $TMPDIR/dumps
 
 # Generate Reading Log/Ratings dumps
@@ -65,19 +69,27 @@ time psql $PSQL_PARAMS --set=upto="$date" -f $SCRIPTS/dump-ratings.sql | gzip -c
 ls -lhR
 
 log "generating the data table: data.txt.gz -- takes approx. 110 minutes..."
-# NOTE: When testing, it is useful to save `data.txt.gz` file to another directory.
-time psql $PSQL_PARAMS -c "copy data to stdout" | gzip -c > data.txt.gz  # TESTING: comment this out
+# In production, we copy the contents of our database into the `data.txt.gz` file.
+# else if we are testing, save a lot of time by using a preexisting `data.txt.gz`.
+if [[ -z $OLDUMP_TESTING ]]; then
+    time psql $PSQL_PARAMS -c "copy data to stdout" | gzip -c > data.txt.gz
+fi
 ls -lhR  # data.txt.gz is 29G
 
 # generate cdump, sort and generate dump
 log "generating $cdump.txt.gz -- takes approx. 500 minutes for 192,000,000+ records..."
+# if $OLDUMP_TESTING has been exported then `oldump.py cdump` will only process a subset.
 time $SCRIPTS/oldump.py cdump data.txt.gz $date | gzip -c > $cdump.txt.gz
 log "generated $cdump.txt.gz"
 ls -lhR  # ol_cdump_2021-11-14.txt.gz is 25G
 
 echo "deleting the data table dump"
 # remove the dump of data table
-rm -f data.txt.gz  # TESTING: comment this out
+# In production, we remove the raw database dump to save disk space.
+# else if we are testing, we keep the raw database dump for subsequent test runs.
+if [[ -z $OLDUMP_TESTING ]]; then
+    rm -f data.txt.gz
+fi
 
 echo "generating the dump -- takes approx. 485 minutes for 173,000,000+ records..."
 time gzip -cd $cdump.txt.gz | python $SCRIPTS/oldump.py sort --tmpdir $TMPDIR | python $SCRIPTS/oldump.py dump | gzip -c > $dump.txt.gz
@@ -106,9 +118,12 @@ function archive_dumps() {
     python /olsystem/bin/uploaditem.py $cdump --nowait --uploader=openlibrary@archive.org
 }
 
+# Only archive if that caller has requested it and we are not testing.
 if [ "$archive" == "--archive" ];
 then
-    archive_dumps
+    if [[ -z $OLDUMP_TESTING ]]; then
+        archive_dumps
+    fi
 fi
